+++ src/main/resources/lessons/httpbasics/documentation/HttpBasics_content1.adoc
