+++ src/main/resources/lessons/httpbasics/documentation/HttpBasics_content2.adoc
