+++ src/main/resources/lessons/httpbasics/documentation/HttpBasics_plan.adoc
