+++ src/main/resources/lessons/httpbasics/html/HttpBasics.html
@@ -6,13 +6,13 @@
         <!-- reuse this lesson-page-wrapper block for each 'page' of content in your lesson -->
 		<!-- include content here, or can be placed in another location. Content will be presented via asciidocs files,
 		which you put in src/main/resources/lessonplans/{lang}/{fileName}.adoc -->
-		<div class="adoc-content" th:replace="doc:lessons/http_basics/documentation/HttpBasics_plan.adoc"></div>
+		<div class="adoc-content" th:replace="doc:lessons/httpbasics/documentation/HttpBasics_plan.adoc"></div>
 	</div>
 
 	<div class="lesson-page-wrapper">
 		<!-- reuse this block for each 'page' of content -->
         <!-- sample ascii doc content for second page -->
-		<div class="adoc-content" th:replace="doc:lessons/http_basics/documentation/HttpBasics_content1.adoc"></div>
+		<div class="adoc-content" th:replace="doc:lessons/httpbasics/documentation/HttpBasics_content1.adoc"></div>
         <!-- if including attack, reuse this section, leave classes in place -->
 		<div class="attack-container">
 			<div class="assignment-success"><i class="fa fa-2 fa-check hidden" aria-hidden="true"></i></div>
@@ -42,7 +42,7 @@
         <!-- reuse this lesson-page-wrapper block for each 'page' of content in your lesson -->
         <!-- include content here. Content will be presented via asciidocs files,
         which you put in src/main/resources/plugin/lessonplans/{lang}/{fileName}.adoc -->
-		<div class="adoc-content" th:replace="doc:lessons/http_basics/documentation/HttpBasics_content2.adoc"></div>
+		<div class="adoc-content" th:replace="doc:lessons/httpbasics/documentation/HttpBasics_content2.adoc"></div>
 		<div class="attack-container">
 			<div class="assignment-success"><i class="fa fa-2 fa-check hidden" aria-hidden="true"></i></div>
 			<!-- using attack-form class on your form, will allow your request to be ajaxified and stay within the display framework for webgoat -->
