+++ src/main/resources/lessons/httpbasics/i18n/WebGoatLabels_de.properties
