+++ src/main/resources/lessons/httpbasics/i18n/WebGoatLabels_nl.properties
